+++ examples/CLI/test.sh
@@ -22,6 +22,7 @@ port="${PORT-20000}"
 # Test data (picked randomly from testnet).
 some_block_hash="970112d640a183b317f79ca9cc4db8cac3f1e263c68fca8c19d14b9fd2041a74"
 some_account_address="33Po4Z5v4DaAHo9Gz9Afc9LRzbZmYikus4Q7gqMaXHtdS17khz"
+some_mod_ref="c14efbca1dcf314c73cc294cbbf1bd63e3906b20d35442943eb92f52e383fc38"
 
 # Build CLI.
 swift build
@@ -42,3 +43,13 @@ cli_path="${dir_path}/concordium-example-client"
 "${cli_path}" --host="${host}" --port="${port}" consensus-info
 "${cli_path}" --host="${host}" --port="${port}" chain-parameters
 "${cli_path}" --host="${host}" --port="${port}" chain-parameters --block="${some_block_hash}"
+"${cli_path}" --host="${host}" --port="${port}" election-info
+"${cli_path}" --host="${host}" --port="${port}" election-info --block="${some_block_hash}"
+"${cli_path}" --host="${host}" --port="${port}" tokenomics-info
+"${cli_path}" --host="${host}" --port="${port}" tokenomics-info --block="${some_block_hash}"
+"${cli_path}" --host="${host}" --port="${port}" finalized-blocks --num-blocks 2
+"${cli_path}" --host="${host}" --port="${port}" module-source --module-ref="${some_mod_ref}"
+"${cli_path}" --host="${host}" --port="${port}" invoke-instance --index 1999 --subindex 0 --entrypoint "voting.view"
+"${cli_path}" --host="${host}" --port="${port}" bakers --block="${some_block_hash}"
+"${cli_path}" --host="${host}" --port="${port}" pool-info --block="${some_block_hash}" --baker-id 9999
+"${cli_path}" --host="${host}" --port="${port}" passive-pool-info --block="${some_block_hash}"