+++ CHANGELOG.md
@@ -16,10 +16,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 #### GRPC client
 - `NodeClient.status` to query transaction status.
 - `SubmittedTransaction` to provide ergonomics for transactions submitted to a node.
-- `NodeClient.finalizedBlocks` to query finalized blocks added from the time the query is made
+- `NodeClient.finalizedBlocks` to query finalized blocks added from the time the query is made.
 - `NodeClient.waitUntilFinalization` helper function to get the summary of a transaction and the block it is finalized into while waiting for it to be finalized.
 - `NodeClient.consensusInfo` for querying the consensus info of the chain.
-- `NodeClient.chainParameters` for querying the the parameters of the chain.
+- `NodeClient.chainParameters` for querying the parameters of the chain.
+- `NodeClient.electionInfo` for querying the election info of the chain containing data regarding active bakers and election of these.
+- `NodeClient.tokenomicsInfo` for querying the tokenomics info of the chain.
+- `NodeClient.source` for querying the the module source corresponding to a module reference.
+- `NodeClient.invokeInstance` for running contract instance entrypoint invocations, returning an invocation result.
+- `NodeClient.bakers` for querying a list of bakers/validators from the node.
+- `NodeClient.poolInfo` for querying the pool info for a baker/validator identified by it's baker ID.
+- `NodeClient.passiveDelegationInfo` for querying the passive delegation pool info of the chain.
 
 ### Changed